requirements.txt
pillow==10.2.0
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2